+++ Mailer/extensions/mailer-module/src/Repositories/AuditLogRepository.php
@@ -4,31 +4,26 @@ declare(strict_types=1);
 namespace Remp\MailerModule\Repositories;
 
 use Nette\Database\Explorer;
-use Nette\Security\IUserStorage;
+use Nette\Security\UserStorage;
 
 class AuditLogRepository extends Repository
 {
     /** @var string */
     protected $tableName = 'audit_logs';
 
-    /** @var IUserStorage */
-    protected $userStorage;
-
     const OPERATION_CREATE = 'create';
     const OPERATION_READ = 'read';
     const OPERATION_UPDATE = 'update';
     const OPERATION_DELETE = 'delete';
 
-    public function __construct(Explorer $database, IUserStorage $userStorage)
+    public function __construct(Explorer $database, protected UserStorage $userStorage)
     {
         parent::__construct($database);
-        $this->database = $database;
-        $this->userStorage = $userStorage;
     }
 
     public function add(string $operation, string $tableName, string $signature, array $data = [])
     {
-        $identity = $this->userStorage->getIdentity();
+        [$isAuthenticated, $identity, $reason] = $this->userStorage->getState();
         $userId = $identity ? $identity->getId() : null;
 
         return $this->insert([