+++ CHANGELOG.md
@@ -33,6 +33,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ### [Mailer]
 
+- **BREAKING**: Updated `nette/security` package removing `IUserStorage` interface.
+    - If you implemented this interface, use `UserStorage` instead.
 - Added index to `mail_job_batch.status` to improve Mailer's workers batch checking performance.
 - Added index to `mail_template_stats` to improve Mailer's dashboard aggregation performance.
 - Allowed cross-origin requests in MailGeneratorFormFactory (opt-in) to support cross-domain submissions if necessary.